Makes the calculator buttons usable.
To acheive this result, I had to comment the code that allow to use
automatic tests. The commented code is located in the calculator.js file.
Precisely the line that contains the 'module.exports' part, as it is not
compatible with the browser execution. Until I find a way to correct this,
I will have to uncomment this line and comment the 'export default' part
just below to run Jest.